Removed obsolete return code: `ResultImportStopped`

+++ concordium-node/src/consensus_ffi/ffi.rs
@@ -454,7 +454,7 @@ extern "C" {
         consensus: *mut consensus_runner,
         import_file_path: *const u8,
         import_file_path_len: i64,
-    ) -> u8;
+    ) -> u64;
     pub fn stopImportingBlocks(consensus: *mut consensus_runner);
 }
 
@@ -871,7 +871,7 @@ impl ConsensusContainer {
         )))
     }
 
-    pub fn import_blocks(&self, import_file_path: &[u8]) -> u8 {
+    pub fn import_blocks(&self, import_file_path: &[u8]) -> u64 {
         let consensus = self.consensus.load(Ordering::SeqCst);
         let len = import_file_path.len();
